# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Invoice API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/invoice/generate": {
            "post": {
                "tags": [
                    "Invoice"
                ],
                "summary": "Generate an invoice from the customer's cart",
                "description": "This endpoint generates an invoice for the customer based on the products in their cart. It calculates the subtotal, applies any discount, adds tax, and returns the total amount for the invoice.",
                "operationId": "2b308879ce570e4448a665a18ffbe327",
                "requestBody": {
                    "description": "Invoice data to generate, including customer ID, tax rate, and optional discount",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tax_rate",
                                    "customer_id"
                                ],
                                "properties": {
                                    "discount": {
                                        "description": "Discount applied to the total amount (optional)",
                                        "type": "number",
                                        "format": "float",
                                        "example": 10
                                    },
                                    "tax_rate": {
                                        "description": "Tax rate to apply to the subtotal (in percentage)",
                                        "type": "number",
                                        "format": "float",
                                        "example": 15
                                    },
                                    "customer_id": {
                                        "description": "ID of the customer whose cart will be used for the invoice",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Invoice generated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invoice generated successfully"
                                        },
                                        "invoice": {
                                            "properties": {
                                                "id": {
                                                    "description": "Invoice ID",
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "customer_name": {
                                                    "description": "Customer's name",
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "customer_email": {
                                                    "description": "Customer's email",
                                                    "type": "string",
                                                    "example": "johndoe@example.com"
                                                },
                                                "customer_address": {
                                                    "description": "Customer's address",
                                                    "type": "string",
                                                    "example": "1234 Elm St."
                                                },
                                                "subtotal": {
                                                    "description": "Subtotal of the cart items",
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 100
                                                },
                                                "discount": {
                                                    "description": "Discount applied",
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 10
                                                },
                                                "tax_rate": {
                                                    "description": "Tax rate applied",
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 15
                                                },
                                                "tax_amount": {
                                                    "description": "Tax amount calculated",
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 13.5
                                                },
                                                "total_amount": {
                                                    "description": "Total amount after discount and tax",
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 103.5
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request - Cart is empty or missing required parameters",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Cart is empty"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No products in cart to generate invoice."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Customer not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Customer not found"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "The customer associated with this cart could not be found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Validation Error"
                                        },
                                        "messages": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string"
                                                }
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Server Error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "An unexpected error occurred while generating the invoice."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}